Moved memory structs into their own headers.
